% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_heights_from_lidar.R
\name{calc_heights_from_lidar}
\alias{calc_heights_from_lidar}
\title{Generate height raster from LiDAR data}
\usage{
calc_heights_from_lidar(
  las = NULL,
  res = 1,
  thresholds = c(0, 2, 5, 10, 20),
  ...
)
}
\arguments{
\item{las}{LiDAR data (a \code{\link[lidR:LAS-class]{lidR::LAS-class()}} or \code{\link[lidR:LAScatalog-class]{lidR::LAScatalog-class()}} object).}

\item{res}{Spatial resolution of the raster}

\item{thresholds}{Set of height thresholds (see \code{\link[lidR:dsm_pitfree]{lidR::dsm_pitfree()}}).}

\item{...}{further arguments to \code{\link[lidR:dsm_pitfree]{lidR::pitfree()}}}
}
\value{
SpatRaster with height data
}
\description{
Generate height raster from LiDAR data
}
\examples{
\dontrun{
lidar.file <- system.file("extdata", "PlazaNueva.laz",
package = "CityShadeMapper", mustWork = TRUE)
lidar <- read_lidar(lidar.file)
heights <- calc_heights_from_lidar(lidar)
}

}
