% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shaded_route.R
\name{calc_shaded_route}
\alias{calc_shaded_route}
\title{Calculate shaded route}
\usage{
calc_shaded_route(origin = NULL, destination = NULL, date = NULL, time = NULL)
}
\arguments{
\item{origin}{origin coordinates.}

\item{destination}{destination coordinates.}

\item{date}{Character. Date in YYYY-MM-DD format.}

\item{time}{Numeric. Hour in local time.}
}
\description{
Calculate shaded route
}
