% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_building}
\alias{example_building}
\title{Single building example}
\usage{
example_building()
}
\value{
A SpatRaster.
}
\description{
An example 9 x 9 meter SpatRaster with a 4 m tall building in the middle.
}
