% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_square}
\alias{example_square}
\title{Square example}
\usage{
example_square()
}
\value{
A SpatRaster.
}
\description{
An example 100 x 100 meter SpatRaster with a 10 m tall rectangular wall and
a 4-m tall tree in the middle.
}
