% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_city_lonlat.R
\name{find_city_lonlat}
\alias{find_city_lonlat}
\title{Find city centre coordinates}
\usage{
find_city_lonlat(ras = NULL)
}
\arguments{
\item{ras}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}} object. Must have a defined CRS}
}
\value{
A list with two elements: lon and lat, in geographic coordinates.
}
\description{
Find the city centre coordinates (longitude & latitude) from the provided raster.
}
\keyword{internal}
