% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sun_position.R
\name{get_sun_position}
\alias{get_sun_position}
\title{Get sun position}
\usage{
get_sun_position(lon = NULL, lat = NULL, date = NULL, hour = NULL)
}
\arguments{
\item{lon}{Longitude (numeric value between -180 and 180)}

\item{lat}{Latitude (numeric value between -90 and 90)}

\item{date}{A character giving the date (e.g. "2021-02-19")}

\item{hour}{Hour of the day. Integer number between 0 and 23, both included.}
}
\value{
A numeric matrix as returned by \code{\link[solartime:computeSunPositionDoyHour]{solartime::computeSunPositionDoyHour()}},
but converted to degrees rather than radians.
}
\description{
Get sun position (declination, elevation, and azimuth) for a given day and time.
}
\examples{
sunpos <- get_sun_position(lon = -5.99, lat = 37.39, date = "2021-02-19", hour = 15)
}
