% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_shademap.R
\name{leaflet_shademap}
\alias{leaflet_shademap}
\title{Make leaflet shade map}
\usage{
leaflet_shademap(
  url.canopy = NULL,
  url.ground = NULL,
  url.cog = NULL,
  band = 1,
  satellite = FALSE,
  opacity = 0.5
)
}
\arguments{
\item{url.canopy}{Character. Url pointing to shade tiles at the canopy level.}

\item{url.ground}{Character. Url pointing to shade tiles at the ground level.}

\item{url.cog}{Character. Url pointing to a cloud-optimised geotiff
(starting with 'https://'). The COG raster must have EPSG:4326 projection.
Currently not implemented.}

\item{band}{Numeric. Band to show in multilayer rasters (only for COG).}

\item{satellite}{Logical. Add satellite images as another layer to the map?}

\item{opacity}{Numeric between 0 and 1. Opacity of the shade layer.}
}
\value{
A leaflet map
}
\description{
Make leaflet shade map
}
\examples{
\dontrun{
leaflet_shademap(url.canopy = "https://mapasdesombra.github.io/Sevilla-jul-canopy/11/{z}/{x}/{y}.png")
}
}
