% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shademap.R
\name{make_shademap}
\alias{make_shademap}
\title{Calculate shade map}
\usage{
make_shademap(
  height.ras = NULL,
  date = NULL,
  hour = NULL,
  type = c("canopy", "ground"),
  cover.ras = NULL,
  zscale = 1,
  omit.nights = TRUE,
  filename = NULL,
  ...
)
}
\arguments{
\item{height.ras}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}} with heights data.
Note \code{height.ras} must have a well defined \code{crs} (see \code{\link[terra:crs]{terra::crs()}}).}

\item{date}{A Date object or character giving the date in YYYY-MM-DD format (e.g. "2021-02-19").
Can be a vector too, e.g. c("2021-02-19", "2021-08-04").
Note different years have no effect on sun position calculations.}

\item{hour}{Hour of the day. Integer number (or numeric vector) between 0 and 23 (both included).}

\item{type}{Character. Either 'canopy' to get illumination at the canopy/roof level, or
'ground' to get illumination at the ground level (i.e. for pedestrians).}

\item{cover.ras}{A SpatRaster containing cover classes (ground, vegetation, buildings...).
See \code{\link[=rasterize_lidar_cover_class]{rasterize_lidar_cover_class()}}.}

\item{zscale}{Default \code{1}. The ratio between the x and y spacing (which are assumed to be equal)
and the z axis. For example, if the elevation is in units of meters and the grid values
are separated by 10 meters, \code{zscale} would be 10.}

\item{omit.nights}{Logical. If TRUE, sun positions will only be returned when it is daytime
(i.e. nighttimes will be omitted)}

\item{filename}{Character. Output filename. Note that if a file already exists
with that name, it will be overwritten.}

\item{...}{further arguments to \code{\link[rayshader:ray_shade]{rayshader::ray_shade()}}}
}
\value{
A (possibly multilayer) SpatRaster object with the intensity of illumination
at each pixel for every date and time
}
\description{
Calculate shade raster for given dates and hours
}
\examples{
\dontrun{
lidar <- PlazaNueva()
heights <- calc_heights_from_lidar(lidar)
shaderas <- make_shademap(heights, date = "2022-10-15", hour = 13)
plot_shademap(shaderas, smooth = TRUE)
shaderas <- make_shademap(heights, date = "2022-10-15", hour = 8:20)
plot_shademap(shaderas, animate = TRUE, smooth = TRUE)
shaderas <- make_shademap(heights, date = "2022-07-15", hour = 8:21)
plot_shademap(shaderas, animate = TRUE, smooth = TRUE)
shaderas <- make_shademap(heights, date = c("2022-07-15", "2022-10-15"), hour = 13)
plot_shademap(shaderas, legend = FALSE)

## Ground-level shade maps require additional raster with cover classes
lidar <- read_lidar(system.file("extdata", "PlazaNueva.laz", package = "CityShadeMapper"))
cover.ras <- rasterize_lidar_cover_class(lidar)
shaderas <- make_shademap(heights, date = "2022-10-15", hour = 13,
  type = "ground", cover.ras = cover.ras)

}
}
