% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shademap.R
\name{make_shademap_ground}
\alias{make_shademap_ground}
\title{Calculate shade map at the ground level}
\usage{
make_shademap_ground(shademap.canopy = NULL, cover.ras = NULL)
}
\arguments{
\item{shademap.canopy}{A SpatRaster with the illumination at the canopy level,
made with \code{\link[=make_shademap]{make_shademap()}}.}

\item{cover.ras}{A SpatRaster containing cover classes (ground, vegetation, buildings...).
See \code{\link[=rasterize_lidar_cover_class]{rasterize_lidar_cover_class()}}.}
}
\value{
A (possibly multilayer) SpatRaster object with the intensity of illumination
at the ground level
}
\description{
Calculate shade map at the ground level
}
\examples{
\dontrun{
lidar <- read_lidar(system.file("extdata", "PlazaNueva.laz", package = "CityShadeMapper"))
cover.ras <- rasterize_lidar_cover_class(lidar)
shaderas.canopy <- make_shademap(heights, date = "2022-10-15", hour = 13)
shaderas.ground <- make_shademap_ground(shaderas.canopy, cover.ras)

## Alternatively, call make_shademap directly:
heights <- calc_heights_from_lidar(lidar)
shaderas.ground <- make_shademap(heights, date = "2022-10-15", hour = 13,
  type = "ground", cover.ras = cover.ras)
}
}
