% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_to_matrix.R
\name{matrix_to_terra}
\alias{matrix_to_terra}
\title{Convert a matrix into a SpatRaster}
\usage{
matrix_to_terra(mat = NULL, crs = "", extent = NULL, ...)
}
\arguments{
\item{mat}{A matrix}

\item{crs}{CRS of the resulting SpatRaster}

\item{extent}{extent}

\item{...}{Further arguments for \code{\link[terra:rast]{terra::rast()}}}
}
\value{
A SpatRaster object
}
\description{
Convert a matrix into a SpatRaster
}
