% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movepoints2nearestcell.R
\name{movepoints2nearestcell}
\alias{movepoints2nearestcell}
\title{Move point locations to the nearest raster cell with data}
\usage{
movepoints2nearestcell(locs = NULL, ras = NULL, layer = 1)
}
\arguments{
\item{locs}{A point \code{sf} object.}

\item{ras}{A SpatRaster.}

\item{layer}{Integer. For multilayer rasters, which layer to use.}
}
\value{
A sf object (with corrected coordinates if move is TRUE).
}
\description{
Move point locations falling in raster cells without data (i.e. NA) to the
nearest raster cell with data.
}
\keyword{internal}
