% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PlazaNueva}
\alias{PlazaNueva}
\title{Plaza Nueva LiDAR data}
\usage{
PlazaNueva()
}
\value{
A LAScatalog.
}
\description{
LiDAR data of Plaza Nueva in Sevilla, Spain.
Data provided by LiDAR-PNOA 2018 CC-BY 4.0 scne.es.
}
