% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shademap.R
\name{plot_shademap}
\alias{plot_shademap}
\title{Plot shade map}
\usage{
plot_shademap(shade.ras = NULL, legend = TRUE, animate = FALSE, ...)
}
\arguments{
\item{shade.ras}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}} object with shade intensity, as produced by \code{\link[=make_shademap]{make_shademap()}}.}

\item{legend}{Logical. Show legend?}

\item{animate}{Logical. Show animation of all the \code{shade.ras} layers?}

\item{...}{Further arguments to \code{\link[terra:plot]{terra::plot()}}, or to \code{\link[terra:animate]{terra::animate()}} if \code{animate} is \code{TRUE}.}
}
\value{
A static or animated plot.
}
\description{
Plot shade map
}
