% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movepoints2nearestcell.R
\name{point_in_cell}
\alias{point_in_cell}
\title{Do points (occurrences) fall within raster cells?}
\usage{
point_in_cell(locs = NULL, ras = NULL, layer = 1)
}
\arguments{
\item{locs}{A point \code{sf} or \code{SpatVector} object with (x, y) coordinates}

\item{ras}{SpatRaster object}

\item{layer}{Integer. For multilayer rasters, which layer to use to compare
with point locations (default = 1).}
}
\value{
A logical vector.
}
\description{
This function examines which points fall within a raster cell with data (not NA).
Returns TRUE for points falling in a raster cell with data, and FALSE otherwise.
}
\keyword{internal}
