% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_to_matrix.R
\name{terra_to_matrix}
\alias{terra_to_matrix}
\title{Convert a SpatRaster to matrix}
\usage{
terra_to_matrix(ras = NULL)
}
\arguments{
\item{ras}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}} object, or a path (character) to a raster file.}
}
\value{
A matrix
}
\description{
Convert a SpatRaster to matrix
}
