% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_heights_from_lidar.R
\name{calc_heights_from_lidar}
\alias{calc_heights_from_lidar}
\title{Calculate height LiDAR}
\usage{
calc_heights_from_lidar(
  las = NULL,
  res = 1,
  thresholds = c(0, 2, 5, 10, 20),
  ...
)
}
\arguments{
\item{las}{Lidar data as LAS or LAScatalog from lidR package format}

\item{res}{Resolution of the raster height}

\item{thresholds}{Set of height thresholds (see lidR package for more information)}

\item{...}{further arguments to \code{\link[lidR:dsm_pitfree]{lidR::pitfree()}}}
}
\value{
SpatRaster object with height LiDAR data
}
\description{
Calculate height LiDAR
}
