% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shade.R
\name{calc_shade}
\alias{calc_shade}
\title{Calculate shades from elevation map}
\usage{
calc_shade(
  height.mat = NULL,
  sun.altitude = NULL,
  sun.angle = NULL,
  zscale = 1,
  crs = "",
  extent = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{height.mat}{Height matrix, as calculated by \code{\link[=terra_to_matrix]{terra_to_matrix()}}.}

\item{sun.altitude}{Sun altitude (angle from the horizon, in degrees)}

\item{sun.angle}{Sun angle in degrees (zero degrees is North, increasing clockwise)}

\item{zscale}{Default \code{1}. The ratio between the x and y spacing (which are assumed to be equal)
and the z axis. For example, if the elevation is in units of meters and the grid values
are separated by 10 meters, \code{zscale} would be 10.}

\item{crs}{coordinate reference system, see \code{\link[terra:crs]{terra::crs()}}}

\item{extent}{raster extent, see \code{\link[terra:ext]{terra::ext()}}}

\item{filename}{Character. Optional. Provide a filename to save the output raster on disk.}

\item{...}{further arguments to \code{\link[rayshader:ray_shade]{rayshader::ray_shade()}}}
}
\value{
A SpatRaster of light intensities, as produced by \code{\link[rayshader:ray_shade]{rayshader::ray_shade()}}
}
\description{
Calculate shades from elevation map
}
\examples{
\dontrun{
lidar.file <- system.file("extdata", "PlazaNueva.laz",
package = "CityShadeMapper", mustWork = TRUE)
lidar <- read_lidar(lidar.file)
heights <- calc_heights_from_lidar(lidar)
library(terra)
height.mat <- CityShadeMapper:::terra_to_matrix(subset(heights, 1))
sun.pos <- get_sun_position(lon = -5.99567, lat = 37.38876, date = "2022-10-18", hour = 15)
shaderas <- calc_shade(height.mat, sun.altitude = sun.pos$elevation,
sun.angle = sun.pos$azimuth,
crs = crs(heights), extent = ext(heights))
plot_shademap(shaderas)
}
}
\keyword{internal}
