% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shade.R
\name{calc_shade}
\alias{calc_shade}
\title{Calculate shades from elevation map}
\usage{
calc_shade(
  height.mat = NULL,
  sun.altitude = NULL,
  sun.angle = NULL,
  crs = "",
  extent = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{height.mat}{Height matrix, as calculated by \code{\link[=terra_to_matrix]{terra_to_matrix()}}.}

\item{sun.altitude}{Sun altitude (angle from the horizon, in degrees)}

\item{sun.angle}{Sun angle in degrees (zero degrees is North, increasing clockwise)}

\item{crs}{coordinate reference system, see \code{\link[terra:crs]{terra::crs()}}}

\item{extent}{raster extent, see \code{\link[terra:ext]{terra::ext()}}}

\item{filename}{Character. Optional. Provide a filename to save the output raster on disk.}

\item{...}{further arguments to \code{\link[rayshader:ray_shade]{rayshader::ray_shade()}}}
}
\value{
A SpatRaster of light intensities, as produced by \code{\link[rayshader:ray_shade]{rayshader::ray_shade()}}
}
\description{
Calculate shades from elevation map
}
