% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shaded_route.R
\name{calc_shaded_route}
\alias{calc_shaded_route}
\title{Calculate shaded route}
\usage{
calc_shaded_route(origin = NULL, dest = NULL, shaderas = NULL)
}
\arguments{
\item{origin}{origin coordinates.}

\item{shaderas}{Shade raster}

\item{destination}{destination coordinates.}
}
\value{
A SpatVector object with the optimal route
}
\description{
Calculate shaded route
}
