% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shaded_route.R
\name{calc_shaded_route}
\alias{calc_shaded_route}
\title{Title}
\usage{
calc_shaded_route(origin = NULL, destination = NULL, date = NULL, time = NULL)
}
\arguments{
\item{time}{}
}
\description{
Title
}
