% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shaded_route.R
\name{calc_shaded_route}
\alias{calc_shaded_route}
\title{Calculate shaded route}
\usage{
calc_shaded_route(origin = NULL, destination = NULL, shaderas = NULL)
}
\arguments{
\item{origin}{A character vector describing an address, an \code{sf} or a \code{SpatVector} object.}

\item{destination}{A character vector describing an address, an \code{sf} or a \code{SpatVector} object.}

\item{shaderas}{Shade raster}
}
\value{
A SpatVector object with the optimal route
}
\description{
Calculate shaded route
}
\examples{
\dontrun{
shaderas <- terra::rast("/vsicurl/https://zenodo.org/record/7213637/files/m7_h13_ground.tif")
shade.route <- calc_shaded_route("Plaza Nueva, Sevilla", "Mateos Gago, Sevilla", shaderas)

library(leaflet)
leaflet(sf::st_as_sf(shade.route)) |>
  leaflet::addWMSTiles(baseUrl = "https://www.ign.es/wms-inspire/ign-base",
  layers = "IGNBaseTodo-nofondo") |>
  leaflet::addTiles(urlTemplate = "https://mapasdesombra.github.io/Sevilla-jul-ground/13/{z}/{x}/{y}.png",
  options = leaflet::tileOptions(minZoom = 15, maxZoom = 18, tms = TRUE, opacity = 0.5)) |>
  addPolylines(weight = 8, opacity = 0.8)

}
}
