% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CityShadeMapper-package.R
\docType{package}
\name{CityShadeMapper-package}
\alias{CityShadeMapper}
\alias{CityShadeMapper-package}
\title{CityShadeMapper: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Francisco Rodríguez-Sánchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
