% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shade_raster.R
\name{get_shade_raster}
\alias{get_shade_raster}
\title{Get shade raster}
\usage{
get_shade_raster(
  height.ras = NULL,
  date = NULL,
  hour = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{height.ras}{A \code{\link[terra:SpatRaster-class]{terra::SpatRaster()}} with heights data}

\item{filename}{Character. Optional. Provide a filename to save the output raster on disk.}

\item{...}{
  Arguments passed on to \code{\link[=calc_shade]{calc_shade}}
  \describe{
    \item{\code{height.mat}}{Height matrix, as calculated by \code{\link[=terra_to_matrix]{terra_to_matrix()}}.}
    \item{\code{sun.altitude}}{Sun altitude (angle from the horizon, in degrees)}
    \item{\code{sun.angle}}{Sun angle in degrees (zero degrees is North, increasing clockwise)}
    \item{\code{crs}}{coordinate reference system, see \code{\link[terra:crs]{terra::crs()}}}
    \item{\code{extent}}{raster extent, see \code{\link[terra:ext]{terra::ext()}}}
  }}
}
\value{
A (possibly multilayer) SpatRaster object with the intensity of shading
at each pixel for every date and time
}
\description{
Get a shade raster for given dates and hours
}
