% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_shademap.R
\name{leaflet_shademap}
\alias{leaflet_shademap}
\title{Make leaflet shade map}
\usage{
leaflet_shademap(url = NULL, orthoimages = FALSE)
}
\arguments{
\item{url}{Character. Url pointing to a cloud-optimised geotiff
(starting with 'https://'). The COG raster must have EPSG:4326 projection.}

\item{orthoimages}{Logical. Add orthoimage layer to the map?}
}
\value{
A leaflet map
}
\description{
Make leaflet shade map
}
\examples{
\dontrun{
leaflet_shademap(url = "https://www.dropbox.com/s/e6l434gmj9v0bpk/cog.tif?dl=1")
}
}
