% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_lidar_cover_class.R
\name{rasterize_lidar_cover_class}
\alias{rasterize_lidar_cover_class}
\title{Rasterize cover classification of lidar points}
\usage{
rasterize_lidar_cover_class(las = NULL, res = 1)
}
\arguments{
\item{las}{Path to a LAS/LAZ object.}

\item{res}{Resolution of the resulting raster.}
}
\value{
A SpatRaster with the classification of cover types:
9 = water, 6 = buildings, 5 = vegetation > 3m, 4 = vegetation between 1 and 3 m high,
3 = vegetation < 1m, 2 = ground.
}
\description{
Rasterize cover classification of lidar points
}
\examples{
\dontrun{
pza <- PlazaNueva()
pza.cover <- rasterize_lidar_cover_class(pza)
}
}
