% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_lidar_cover_class.R
\name{rasterize_lidar_cover_class}
\alias{rasterize_lidar_cover_class}
\title{Get raster of cover classification from lidar points}
\usage{
rasterize_lidar_cover_class(
  las = NULL,
  res = 1,
  fill.holes = TRUE,
  filename = NULL
)
}
\arguments{
\item{las}{A \code{\link[lidR:LAScatalog-class]{lidR::LAScatalog-class()}} object, or a character vector with
paths to LAS/LAZ objects.}

\item{res}{Resolution of the resulting raster.}

\item{fill.holes}{Logical. Try to fill holes in lidar point classification.}

\item{filename}{Character. Output filename. Note that if a file already exists
with that name, it will be overwritten.}
}
\value{
A SpatRaster with the classification of cover types:
2 = ground (including low vegetation < 1m)
4 = high vegetation (> 1m)
6 = buildings
9 = water
and NA values.
Note that points classified as bridges (class 17) will be reclassified as ground.
}
\description{
Get raster of cover classification from lidar points
}
\examples{
\dontrun{
pza <- system.file("extdata", "PlazaNueva.laz", package = "CityShadeMapper")
pza.cover <- rasterize_lidar_cover_class(pza)
}
}
