% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cog.R
\name{save_cog}
\alias{save_cog}
\title{Save raster as Cloud Optimised GeoTiff}
\usage{
save_cog(ras = NULL, filename = NULL)
}
\arguments{
\item{ras}{A SpatRaster (e.g. a shade raster as produced by \code{\link[=make_shademap]{make_shademap()}}).}

\item{filename}{Character. Output filename. Must end with '.tif' extension.}
}
\value{
A cloud optimised geotiff saved on disk.
}
\description{
Requires GDAL >= 3.1
}
\keyword{internal}
